♻️ Refactor consume_queue into dailyerosion
